clean code清理：表达式中不要混用有符号数和无符号数
# message auto-generated for no-merge-commit merge:
merge clean-code into master

clean code清理：表达式中不要混用有符号数和无符号数  

Created-by: msprofDeveloper

Match-id-47a38415f9b9f30f57843673bad526ca8430cb9a

+++ analysis/csrc/domain/data_process/ai_task/msproftx_device_processor.cpp
@@ -109,7 +109,7 @@ std::vector<MsprofTxDeviceData> MsprofTxDeviceProcessor::FormatData(
         return processedData;
     }
     MsprofTxDeviceData data;
-    double start;
+    uint64_t start;
     data.deviceId = deviceId;
     std::sort(oriData.begin(), oriData.end(), [](TxDeviceData &lData, TxDeviceData rData) {
         return std::get<4>(lData) < std::get<4>(rData); // 第4位为timestamp
